add ThreadPage/ThreadHeader
