+++ app/image/[image_id]/layout.tsx
@@ -2,7 +2,7 @@ import { ThreadHeader, ThreadPage } from '@/components/pages/thread/ThreadPage'
 import { getConvexSiteUrl } from '@/lib/utils'
 
 export async function generateMetadata({ params }: { params: { thread_id: string } }) {
-  const response = await fetch(`${getConvexSiteUrl()}/page?route=chat&id=${params.thread_id}`)
+  const response = await fetch(`${getConvexSiteUrl()}/page?route=image&id=${params.thread_id}`)
   if (!response.ok) return {}
 
   const data = await response.json()
@@ -21,7 +21,7 @@ export default function Layout({
 }) {
   return (
     <ThreadPage>
-      <ThreadHeader thread_id={params.thread_id} />
+      <ThreadHeader thread_id={''} />
       {children}
     </ThreadPage>
   )
