--- app/images/[thread_id]/@modal/(...)image/[image_id]/default.tsx
@@ -1,3 +1,0 @@
-export default function Default() {
-  return null
-}
